requirements.txt

pandas
numpy
joblib
scikit-learn
tensorflow
matplotlib
fastapi
pydantic
scapy
pymongo
python-dotenv
uvicorn
